# docs/application-openapi.yaml
openapi: 3.1.0
info: {title: Tailors SaaS Application API, version: 1.0.0, description: Tenant-scoped web application API. The dedicated voice contract remains in openapi.yaml.}
servers: [{url: https://api.tailors.app/api/v1}]
security: [{sanctumSession: []}]
paths:
...
  /customers/{customer}: {patch: {operationId: updateCustomer, parameters: [{$ref: '#/components/parameters/CustomerId'}], responses: {'200': {$ref: '#/components/responses/Success'}, '422': {$ref: '#/components/responses/Validation'}}}, delete: {operationId: archiveCustomer, parameters: [{$ref: '#/components/parameters/CustomerId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
  /customers/{customer}/restore: {post: {operationId: restoreCustomer, parameters: [{$ref: '#/components/parameters/CustomerId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
...
  /orders/{order}: {get: {operationId: showOrder, parameters: [{$ref: '#/components/parameters/OrderId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}, delete: {operationId: archiveOrder, parameters: [{$ref: '#/components/parameters/OrderId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
  /orders/{order}/restore: {post: {operationId: restoreOrder, parameters: [{$ref: '#/components/parameters/OrderId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
...
  /orders/{order}/assign: {post: {operationId: assignOrder, parameters: [{$ref: '#/components/parameters/OrderId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
  /orders/{order}/status: {patch: {operationId: updateOrderStatus, parameters: [{$ref: '#/components/parameters/OrderId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
  /orders/{order}/payments: {post: {operationId: payOrder, parameters: [{$ref: '#/components/parameters/OrderId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
...
  /garments/{garment}: {patch: {operationId: updateGarment, parameters: [{$ref: '#/components/parameters/GarmentId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
...
  /garments/{garment}/parts/{part}: {delete: {operationId: deleteGarmentPart, parameters: [{$ref: '#/components/parameters/GarmentId'}, {name: part, in: path, required: true, schema: {type: string}}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
...
  /employees/{employee}: {patch: {operationId: updateEmployee, parameters: [{$ref: '#/components/parameters/EmployeeId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
...
components:
  securitySchemes:
    sanctumSession: {type: apiKey, in: cookie, name: laravel_session, description: Obtain with /sanctum/csrf-cookie; mutations require X-XSRF-TOKEN.}
  parameters:
...
    CustomerId: {name: customer, in: path, required: true, schema: {type: string, minLength: 26, maxLength: 26}}
    OrderId: {name: order, in: path, required: true, schema: {type: string, minLength: 26, maxLength: 26}}
    GarmentId: {name: garment, in: path, required: true, schema: {type: string, minLength: 26, maxLength: 26}}
    EmployeeId: {name: employee, in: path, required: true, schema: {type: string, minLength: 26, maxLength: 26}}
...
  schemas:
...
    Error: {type: object, required: [code, message], properties: {code: {type: string}, field: {type: [string, 'null']}, message: {type: string}}}
    Envelope: {type: object, required: [data, meta, errors], properties: {data: {}, meta: {type: object, additionalProperties: true}, errors: {type: array, items: {$ref: '#/components/schemas/Error'}}}}
...
  responses:
    Success: {description: Successful envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
    Validation: {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}